pyproject.toml
[tool.poetry]
name = "neat-ai-assistant"
version = "0.0.1"
description = "An AI assistant that intellegently handles your data."
authors = ["Niklas Finken <niklasfinken@me.com>"]
license = "MIT"
readme = "README.md"
packages = [{from = "src", include = "neat_ai_assistant"}]


[tool.poetry.dependencies]
python = "^3.11.4"
python-dotenv = "^1.0.0"
duckduckgo-search = "^3.8.3"
pydantic = "^2.0.2"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
sse-starlette = "^1.6.1"
tiktoken = "^0.4.0"
geopy = "^2.3.0"
bs4 = "^0.0.1"
openai = "^1.3.5"


[tool.poetry.dev-dependencies]
mypy = "^1.6.1"
types-beautifulsoup4 = "^4.12.0.7"
types-requests = "^2.31.0.10"
pytest = "*"
black = "*"
pre-commit = "*"

[tool.mypy]
files = "src,tests"
strict = "True"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
